基本完成用户中心的后端接口,以及与前端的连接(还有安全设置,包括修改密码,邮箱,用户名等还未实现,以及头像的读取也没实现) 操作人:孙哲文

+++ it-module-common/src/main/java/com/alikeyou/itmodulecommon/constant/LoginConstant.java
@@ -15,7 +15,7 @@ public class LoginConstant {
     private static String password;
     private static String email;
     private static Long userId; // 用户唯一ID
-    private static Long roleId; // 角色身份ID
+    private static Integer roleId; // 角色身份ID
     
     //setter方法
     public static void setUsername(String username) {
@@ -34,7 +34,7 @@ public class LoginConstant {
         LoginConstant.userId = userId;
     }
     
-    public static void setRoleId(Long roleId) {
+    public static void setRoleId(Integer roleId) {
         LoginConstant.roleId = roleId;
     }
     
@@ -55,7 +55,7 @@ public class LoginConstant {
         return userId;
     }
     
-    public static Long getRoleId() {
+    public static Integer getRoleId() {
         return roleId;
     }
